requirements.txt
requests~=2.31.0

environs~=11.0.0
python-telegram-bot==13.15
redis==3.2.1
